# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CorpChat - Internal Employee Messaging Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with employee credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /cico/status/{employeeId}:
    get:
      operationId: getCicoStatus
      tags: [cico]
      summary: Get CICO status for a specific employee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CICO status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CicoStatus"
  /cico/checkin:
    post:
      operationId: cicoCheckIn
      tags: [cico]
      summary: Manual check-in (sync or simulate CICO)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CicoCheckInRequest"
      responses:
        "200":
          description: Check-in recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CicoStatus"
  /cico/checkout:
    post:
      operationId: cicoCheckOut
      tags: [cico]
      summary: Manual check-out (sync or simulate CICO)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CicoCheckOutRequest"
      responses:
        "200":
          description: Check-out recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CicoStatus"

  # Conversations
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [conversations]
      summary: List company-wide announcements
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Announcements list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementsListResponse"

    post:
      operationId: createAnnouncement
      tags: [conversations]
      summary: Create announcement (admin/manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementRequest"
      responses:
        "201":
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  # Audit Logs
  /audit/logs:
    get:
      operationId: getAuditLogs
      tags: [audit]
      summary: Get audit logs (admin only)
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: conversationId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogsListResponse"
  /audit/conversations/{conversationId}/export:
    get:
      operationId: exportConversation
      tags: [audit]
      summary: Export conversation history (admin only)
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        "200":
          description: Exported conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        employeeId:
          type: string
        password:
          type: string
      required:
        - employeeId
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, manager, employee]
        isActive:
          type: boolean
        cicoStatus:
          $ref: "#/components/schemas/CicoStatus"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - name
        - email
        - role
        - isActive
        - createdAt
...
    CicoStatus:
      type: object
      properties:
        employeeId:
          type: string
        userId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [present, break, wfh, absent, off]
        checkInTime:
          type: string
          format: date-time
          nullable: true
        checkOutTime:
          type: string
          format: date-time
          nullable: true
        location:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required:
        - employeeId
        - status
        - updatedAt
...
    CicoCheckInRequest:
      type: object
      properties:
        employeeId:
          type: string
        location:
          type: string
        type:
          type: string
          enum: [office, wfh]
      required:
        - employeeId
        - type
    CicoCheckOutRequest:
      type: object
      properties:
        employeeId:
          type: string
      required:
        - employeeId
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        author:
          $ref: "#/components/schemas/User"
          nullable: true
        isPinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - authorId
        - isPinned
        - createdAt
    AnnouncementsListResponse:
      type: object
      properties:
        announcements:
          type: array
          items:
            $ref: "#/components/schemas/Announcement"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - announcements
        - total
        - page
        - limit
    CreateAnnouncementRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
      required:
        - title
        - content
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
          nullable: true
        details:
          type: object
          additionalProperties: true
          nullable: true
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - entityType
        - createdAt
    AuditLogsListResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - logs
        - total
        - page
        - limit
    ExportResponse:
      type: object
      properties:
        data:
          type: string
        format:
          type: string
        filename:
          type: string
      required:
        - data
        - format
        - filename
...